AI Branch created
The GameState is so far mostly for the AI Player class to be able to see whether it is in the bidding stage or not.
The AIPlayer class is not yet the complete class, I am still transferring the old class I had been working on to the new project, as when Bitdefender saw my QT as ransom-ware it seems like it messed with my old project.